Semipresencial

+++ app/Http/Controllers/PokemonController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Coach;
 use App\Models\Pokemon;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Gate;
 
 class PokemonController extends Controller
 {
@@ -17,6 +17,8 @@ class PokemonController extends Controller
     
     public function create()
     {
+        Gate::authorize('create', Pokemon::class); 
+
         $coaches = Coach::all();
         return view('pokemons.create', compact('coaches'));
     }
@@ -29,6 +31,8 @@ class PokemonController extends Controller
     
     public function edit($id)
     {
+        Gate::authorize('uptade', Pokemon::class);
+
         $pokemon = Pokemon::findOrFail($id);
         return view('pokemons.edit', compact('pokemon'));
     }
@@ -42,6 +46,8 @@ class PokemonController extends Controller
     
     public function destroy($id)
     {
+        Gate::authorize('delete', Pokemon::class);
+
         $pokemon = Pokemon::findOrFail($id);
         $pokemon->delete();
         return redirect('pokemons')->with('success', 'pokemon deleted successfully.');
